+++ Net_Core/Web.Angular/projects/gw-lib/core/function/src/c/function-details/function-details.component.ts
@@ -117,6 +117,7 @@ export class FunctionDetailsComponent extends BaseDetailComponent implements IBa
 	public derivedViewGroups: Array<string> = [];
 
 	pickListTableContentsBackground = this.clientChoices.alternatingRowBackColor;
+	pickListTableContentsFont = this.clientChoices.headerForeColor;
 	pickListTableHeaderBackground = this.clientChoices.rowBackColor;
 
 	selectedFunctionType: number = 1;